Refactor: implement ReviewGenerator and refactor by extracting method
