+++ src/main/java/com/project/Project/service/building/impl/BuildingServiceImpl.java
@@ -9,8 +9,8 @@ import com.project.Project.repository.building.BuildingRepository;
 import com.project.Project.repository.projection.building.OnlyBuildingIdAndCoord;
 import com.project.Project.repository.review.ReviewRepository;
 import com.project.Project.repository.room.RoomRepository;
+import com.project.Project.service.building.BuildingGenerator;
 import com.project.Project.service.building.BuildingService;
-import com.project.Project.util.component.BuildingGenerator;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
