+++ src/test/java/com/project/Project/service/BuildingServiceTest.java
@@ -5,8 +5,8 @@ import com.project.Project.domain.building.Building;
 import com.project.Project.domain.embedded.Address;
 import com.project.Project.repository.building.BuildingCustomRepository;
 import com.project.Project.repository.building.BuildingRepository;
+import com.project.Project.service.building.BuildingGenerator;
 import com.project.Project.service.building.impl.BuildingServiceImpl;
-import com.project.Project.util.component.BuildingGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
